+++ spec/no_data_bag_match_spec.rb
@@ -17,7 +17,7 @@ platforms = [
 ]
 
 platforms.each { |i| i.each_pair do |p,v|
-  describe 'ssmtp::default' do
+  describe 'ssmtp2::default' do
     let (:chef_run) { ChefSpec::Runner.new(platform:p, version:v, :log_level => :info) do |node|
       Chef::Log.debug(sprintf("#### FILE: %s  PLATFORM: %s  VERSION: %s ####", ::File.basename(__FILE__), p, v))