+++ CHANGELOG.md
@@ -1,6 +1,6 @@
-ssmtp CHANGELOG
+ssmtp2 CHANGELOG
 ===============
 
 0.1.0
 -----
-- [michael.m.morris@gmail.com] - Initial release of ssmtp cookbook
+- [michael.m.morris@gmail.com] - Initial release of ssmtp2 cookbook